requirements.txt
google-adk==1.19.0
google-genai==1.52.0
python-dotenv==1.2.1